workaround: use string_escape=release for nvme symlink rules
The udev default is to not replace whitespace in SYMLINK values,
which (since the SYMLINK list is just a whitespace-separated string)
breaks creation of symlinks for NVMe devices with whitespace in their
model or serial numbers/strings.  Instead of fixing the bug in udev,
which would change the default for all udev rules, this only works
around the bug by updating only the NVMe rules to use the
string_escape=replace rule option, which causes whitespace replacement.
This will not conflict when the bug is eventually fixed, since after
the real bugfix the default behavior will match the
string_escape=replace behavior.