test: add script to convert sys/ into sys-script.sh
Instead of keeping all sys/ nodes in a tarball, use a script
"sys-script.sh" to create all the sys/ entries.

This adds scripts to create that initial "sys-script.sh" script, using
an existing sys/ directory, created from the sys.tar.xz contents.
The "sys-script.sh" can then be edited or recreated later, when any sys/
files are added or modified; the change will be only a patch to the
"sys-script.sh" script in git, instead of forcing git to store a new
binary tarball.