logind: backport v238/v239 fixes for handling DRM devices
These chages introduce all the fixes that correct handling of open fd's
realated to the DRM devices, as used by for example NVIDIA GPUs. This backport
includes some refactorings, corrections, and comment updates. This to insure
that correct history is preserved, code comments match reality, and to easy
backporting logind fixes in the future SRUs.

LP: #1777099