test/test-json: split up test_build() into multiple functions to avoid stack overflow
The JSON_BUILD macros eat up a huge amount of stack when there are a lot
of them in a single function; this was causing build failures because
some of the values overflowed the stack.

Splitting the function up into individual functions eases the pressure
on the stack and avoids the overflow.