network: if !ipv6ll, enumerate ipv6 tentative before dropping foreign addrs
The kernel will create an ipv6ll tentative address immediately when an
interface is raised if the addr_gen_mode is not disabled; and, the kernel
does not notify netlink listeners about any tentative addresses.  So it's
possible for an interface to contain tentative ipv6 link-local address(es)
that networkd doesn't know about when all foreign addresses are dropped.

In this case, networkd is later notified about the new ipv6ll address(es)
after they finish DAD and are no longer tentative; but since that's after
networkd has already dropped foreign addresses, they are incorrectly left
on the interface.

In the call to drop foreign addresses, if ipv6ll is disabled, enumerate
all ipv6 tentative address(es) for the link, so they will be in the
addresses_foreign list and correctly dropped.