network: only honor configure_without_carrier if link never had carrier
The configure_without_carrier option controls if the link is configured
when networkd starts, even if it doesn't have carrier.  There is a separate
option, ignore_carrier_loss, to prevent dropping configuration when the
link loses carrier.  However, if configure_without_carrier is set and
the carrier is lost during configuration of the link, it should only
continue configuration if ignore_carrier_loss is also set.